pyproject.toml
[project]
name = "obrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "debugpy>=1.8.19",
    "fastapi>=0.128.0",
    "google-genai>=1.0.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "sqlmodel>=0.0.16",
    "sentence-transformers>=3.0.0",
    "rank-bm25>=0.2.2",
]
